song data class moved to its own file
