+++ Assets/Script/Serialization/XboxSongData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b6f9d337bdb7874ea66b1660921d6f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
